refactor(domain): reorganize core into Screaming Architecture folders
