+++ core/Answer/answerConstants.ts
@@ -1,0 +1,5 @@
+export const ANSWER_EFFECT = {
+  CORRECT: 'correct',
+  INCORRECT: 'incorrect',
+  NONE: 'none',
+} as const
