+++ core/Game/GameConfig.ts
@@ -1,0 +1,4 @@
+export interface GameConfig {
+  readonly maxErrorsAllowed: number
+  readonly wordsPerGame: number
+}
