+++ core/Game/gameConstants.ts
@@ -1,0 +1,9 @@
+export const GAME_STATUS = {
+  ACTIVE: 'active',
+  FINISHED: 'finished',
+} as const
+
+export const GAME_RESULT = {
+  WON: 'won',
+  LOST: 'lost',
+} as const
