+++ core/Game/logic/finishedGame.ts
@@ -1,0 +1,24 @@
+import { GAME_STATUS } from '@core/Game/gameConstants'
+import type { Game, GameResult } from '@core/Game/Game'
+
+export type FinishedGamePatch = Pick<
+  Game,
+  | 'remainingWords'
+  | 'correctAnswers'
+  | 'wrongAnswers'
+  | 'currentCard'
+  | 'result'
+> & { readonly result: GameResult }
+
+export type ToFinishedGameParams = {
+  readonly base: Game
+  readonly patch: FinishedGamePatch
+}
+
+export function toFinishedGame({ base, patch }: ToFinishedGameParams): Game {
+  return {
+    ...base,
+    ...patch,
+    status: GAME_STATUS.FINISHED,
+  }
+}
