+++ core/Severity/severityConstants.ts
@@ -1,0 +1,6 @@
+export const ERROR_SEVERITY = {
+  GREEN: 'green',
+  YELLOW: 'yellow',
+  ORANGE: 'orange',
+  RED: 'red',
+} as const
