+++ core/Word/Word.ts
@@ -1,0 +1,9 @@
+export type WordGroup = readonly [string, string, string]
+
+export type WordsCollection = readonly WordGroup[]
+
+export interface Word {
+  readonly correctOption: string
+  readonly wrongOptions: [string, string]
+  readonly audioText: string
+}
