+++ core/WordCard/WordCard.ts
@@ -1,0 +1,9 @@
+import type { Word } from '@core/Word/Word'
+
+export interface WordCard {
+  readonly word: Word
+  readonly shuffledOptions: string[]
+  readonly hasBeenAnswered: boolean
+  readonly isCorrect: boolean | null
+  readonly selectedOption: string | null
+}
