+++ core/shared/randomTypes.ts
@@ -1,0 +1,3 @@
+export type RandomIntFn = (maxExclusive: number) => number
+
+export type ShuffleFn = <T>(items: readonly T[]) => T[]
